+++ src/XPike.Caching.Redis/RedisCachingConnectionProvider.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Collections.Concurrent;
 using System.Threading.Tasks;
+using XPike.Configuration;
 using XPike.Logging;
 using XPike.Redis;
-using XPike.Settings;
 
 namespace XPike.Caching.Redis
 {
@@ -15,12 +15,12 @@ namespace XPike.Caching.Redis
         private readonly ConcurrentDictionary<string, IRedisCachingProvider> _connections = new ConcurrentDictionary<string, IRedisCachingProvider>();
 
         private readonly IRedisConnectionProvider _connectionProvider;
-        private readonly ISettings<RedisCachingSettings> _settings;
+        private readonly IConfig<RedisCachingConfig> _config;
         private readonly ILogService _logService;
 
-        public RedisCachingConnectionProvider(ISettings<RedisCachingSettings> settings, ILogService logService, IRedisConnectionProvider connectionProvider)
+        public RedisCachingConnectionProvider(IConfig<RedisCachingConfig> config, ILogService logService, IRedisConnectionProvider connectionProvider)
         {
-            _settings = settings;
+            _config = config;
             _logService = logService;
             _connectionProvider = connectionProvider;
         }
@@ -28,30 +28,30 @@ namespace XPike.Caching.Redis
         public async Task<ICachingProvider> GetConnectionAsync(string connectionName)
         {
             if (connectionName == null)
-                return await GetConnectionAsync(DEFAULT_CONNECTION_NAME);
+                return await GetConnectionAsync(DEFAULT_CONNECTION_NAME).ConfigureAwait(false);
 
             if (_connections.TryGetValue(connectionName, out var connection))
                 return connection;
 
-            if (!_settings.Value.Connections.ContainsKey(connectionName))
+            if (!_config.CurrentValue.Connections.ContainsKey(connectionName))
             {
                 if (connectionName == DEFAULT_CONNECTION_NAME)
                     throw new InvalidOperationException("No configuration found for default Redis caching connection!");
 
                 _logService.Warn($"No configuration found for Redis Caching connection '{connectionName}' - using default connection.");
                 
-                connection = await GetConnectionAsync(DEFAULT_CONNECTION_NAME) as IRedisCachingProvider;
+                connection = await GetConnectionAsync(DEFAULT_CONNECTION_NAME).ConfigureAwait(false) as IRedisCachingProvider;
             }
             else
             {
                 connection = new RedisCachingProvider(connectionName,
-                    _settings,
+                    _config,
                     _logService,
-                    await _connectionProvider.GetConnectionAsync(_settings.Value
+                    await _connectionProvider.GetConnectionAsync(_config.CurrentValue
                         .Connections[connectionName]
-                        .RedisConnectionName));
+                        .RedisConnectionName).ConfigureAwait(false));
 
-                if (!await connection.ConnectAsync())
+                if (!await connection.ConnectAsync().ConfigureAwait(false))
                     _logService.Warn($"Failed to connect to Redis Cache '{connectionName}': Provider returned false!");
             }