+++ src/XPike.Caching/Hybrid/HybridCachingProvider.cs
@@ -18,30 +18,32 @@ namespace XPike.Caching.Hybrid
             _connectionName = connectionName;
         }
 
-        protected async Task<IList<TResult>> ParallelExecuteAsync<TResult>(Func<ICachingProvider, Task<TResult>> asyncOperation) =>
+        protected async Task<IList<TResult>> ParallelExecuteAsync<TResult>(
+            Func<ICachingProvider, Task<TResult>> asyncOperation) =>
             await Task.WhenAll(_providers.Select(async x =>
-                await asyncOperation(await x.GetConnectionAsync(_connectionName))));
+                await asyncOperation(await x.GetConnectionAsync(_connectionName).ConfigureAwait(false))
+                    .ConfigureAwait(false))).ConfigureAwait(false);
 
         protected async Task<IList<ICachingProvider>> GetProvidersAsync() =>
-            (await Task.WhenAll(_providers.Select(x => x.GetConnectionAsync(_connectionName))))
+            (await Task.WhenAll(_providers.Select(x => x.GetConnectionAsync(_connectionName))).ConfigureAwait(false))
             .ToList();
 
         public async Task<bool> InvalidateAsync(string key) =>
-            (await ParallelExecuteAsync<bool>(provider => provider.InvalidateAsync(key)))
+            (await ParallelExecuteAsync<bool>(provider => provider.InvalidateAsync(key)).ConfigureAwait(false))
             .All(x => x);
 
         public async Task<ICachedItem<TItem>> GetItemAsync<TItem>(string key, TimeSpan? timeout = null, CancellationToken? ct = null)
             where TItem : class
         {
             var previousProviders = new List<ICachingProvider>();
-            var providers = await GetProvidersAsync();
+            var providers = await GetProvidersAsync().ConfigureAwait(false);
 
             ICachedItem<TItem> item = null;
             ICachedItem<TItem> bestMatch = null;
 
             foreach (var provider in providers)
             {
-                var found = await provider.GetItemAsync<TItem>(key, timeout, ct);
+                var found = await provider.GetItemAsync<TItem>(key, timeout, ct).ConfigureAwait(false);
 
                 if (!found.IsStale)
                 {
@@ -58,8 +60,9 @@ namespace XPike.Caching.Hybrid
             // NOTE: Intentional fire-and-forget.  We don't want to wait for cache "percolation" to complete.
 #pragma warning disable 4014
             if (item != null)
-                Task.Run(async () =>
-                    await Task.WhenAll(previousProviders.Select(x => x.SetItemAsync(key, item, timeout, ct))));
+                _ = Task.Run(async () =>
+                    await Task.WhenAll(previousProviders.Select(x => x.SetItemAsync(key, item, timeout, ct)))
+                        .ConfigureAwait(false));
 #pragma warning restore 4014
 
             return item ?? bestMatch;
@@ -67,7 +70,7 @@ namespace XPike.Caching.Hybrid
 
         public async Task<bool> SetItemAsync<TItem>(string key, ICachedItem<TItem> item, TimeSpan? timeout = null, CancellationToken? ct = null)
             where TItem : class =>
-            (await ParallelExecuteAsync(provider => provider.SetItemAsync(key, item, timeout, ct)))
+            (await ParallelExecuteAsync(provider => provider.SetItemAsync(key, item, timeout, ct)).ConfigureAwait(false))
             .All(x => x);
     }
 }